+++ tests/magnetometer.rs
@@ -5,11 +5,12 @@ use crate::common::{
     MAG_ADDR,
 };
 use embedded_hal_mock::{
+    delay::MockNoop as Delay,
     i2c::Transaction as I2cTrans,
     pin::{Mock as PinMock, State as PinState, Transaction as PinTrans},
     spi::Transaction as SpiTrans,
 };
-use lsm303agr::MagOutputDataRate as ODR;
+use lsm303agr::{MagMode, MagOutputDataRate as ODR};
 
 macro_rules! set_mag_odr {
     ($name:ident, $hz:ident, $value:expr) => {
@@ -19,7 +20,7 @@ macro_rules! set_mag_odr {
                 MAG_ADDR,
                 vec![Register::CFG_REG_A_M, $value | DEFAULT_CFG_REG_A_M],
             )]);
-            sensor.set_mag_odr(ODR::$hz).unwrap();
+            sensor.set_mag_odr(&mut Delay, ODR::$hz).unwrap();
             destroy_i2c(sensor);
         }
     };
@@ -28,6 +29,33 @@ set_mag_odr!(set_mag_odr_hz10, Hz10, 0);
 set_mag_odr!(set_mag_odr_hz20, Hz20, 1 << 2);
 set_mag_odr!(set_mag_odr_hz50, Hz50, 2 << 2);
 set_mag_odr!(set_mag_odr_hz100, Hz100, 3 << 2);
+
+#[test]
+fn can_change_mode() {
+    let mut sensor = new_i2c(&[
+        // Set low-power mode
+        I2cTrans::write(
+            MAG_ADDR,
+            vec![Register::CFG_REG_A_M, DEFAULT_CFG_REG_A_M | 0b00010000],
+        ),
+        // Set high-resolution mode
+        I2cTrans::write(
+            MAG_ADDR,
+            vec![Register::CFG_REG_A_M, DEFAULT_CFG_REG_A_M | 0b00000000],
+        ),
+    ]);
+    assert_eq!(sensor.get_mag_mode(), MagMode::HighResolution);
+
+    sensor.set_mag_mode(&mut Delay, MagMode::LowPower).unwrap();
+    assert_eq!(sensor.get_mag_mode(), MagMode::LowPower);
+
+    sensor
+        .set_mag_mode(&mut Delay, MagMode::HighResolution)
+        .unwrap();
+    assert_eq!(sensor.get_mag_mode(), MagMode::HighResolution);
+
+    destroy_i2c(sensor);
+}
 
 macro_rules! assert_eq_xyz_nt {
     ($data:expr) => {{